Make querying consistent by normalizing names.
Previously there was no canonicalization of the queries coming into
rainsd. This meant that there were cache misses and spurious
redirections to non-existing servers (because proper delegation is
not yet supported). One such example would be querying for "ch."
when the actual zone in the cache was "ch".

This commit fixes this issue by normalizing all queries to end with
a dot ('.'). This makes sense because the dot signifies the end of
the label and indicates that it is ending under the root zone. It
should be discussed how this relates to the context which is also
part of the query.

BUG=https://github.com/netsec-ethz/rains/issues/78