Formulate the cache key using normalized names.
There were some edge cases when querying the cache that was affected
by normalizing all domain names to end with a dot. This change alters
the way the cache key is joined to behave rationally when querying the
root zone (e.g. not having ".." when trying to concatenate the name
and zone together).